Fix updating neighbour blocks Fix flowerpot saving. Fixes #119

+++ gomint-server/src/main/java/io/gomint/server/world/anvil/tileentity/v1_8/FlowerPotConverter.java
@@ -59,7 +59,7 @@ public class FlowerPotConverter extends BasisConverter<FlowerPotTileEntity> {
 
         // Store the item
         compound.addValue( "Item", MaterialMagicNumbers.newIdFromValue( entity.getHoldingItem().getMaterial() ) );
-        compound.addValue( "Data", entity.getHoldingItem().getData() );
+        compound.addValue( "Data", (int) entity.getHoldingItem().getData() );
     }
 
 }
